Network used to obtain the best results with MobileNetV2 (0.88222 on test)
Network used to obtain the best results with MobileNetV2 (0.88222 on test). After many tests the impression is that the network is good even if it has less parameters than other, more complex, networks. By the way the fact that increasing the complexity of the fully connected part we achieved better results induce the belief that a more complex architecture would perform better.